Implemented more tests and refactored AVAudioPCMBuffer extension.
Implemented PitchYinFFT test. Added stubs for FrameCutter algorithm tests to try and diagnose why PitchYinFFT fails the real case assertions. It seems to be that what should be an all-zero first frame is not so. Refactored the `mono` derived property in the AVAudioPCMBuffer extension into a function that mimics MonoLoader and added a test that takes its values from test_monoloader.py.

Signed-off-by: Jason Cardwell <cardwell.jason@gmail.com>